Added Database; Removed WeatherForecastController; Added Classes
Database should now be set up and working; Removed WFC because it is not needed for the project; Added basic classes. Can now create, get, delete Classes. Just includes an Id, Class name, Description of the class, and who owns it. There are not any classes in the Database at this moment